Refactoring day 01
