+++ 2019/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2024"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+fs = "0.0.5"
 regex = "1.11.1"
