+++ 2019/tests/day01_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day01;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(3339288, day01::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(5006064, day01::part2());
+    }
+}
